@@ -8,26 +8,30 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^7.3",
-        "fruitcake/laravel-cors": "^2.0",
-        "fideloper/proxy": "^4.2",
-        "guzzlehttp/guzzle": "^7.0.1",
-        "laravel/framework": "^8.0",
+        "php": "^7.3|^8.0",
         "benjamincrozat/laravel-dropbox-driver": "^3.0",
+        "fideloper/proxy": "^4.4",
+        "fruitcake/laravel-cors": "^2.0",
         "google/apiclient": "^2.9",
-        "laravel/tinker": "^2.0",
+        "guzzlehttp/guzzle": "^7.0.1",
+        "laravel/framework": "^8.12",
+        "laravel/helpers": "^1.4",
+        "laravel/tinker": "^2.5",
         "nao-pon/flysystem-google-drive": "~1.1",
+        "pbmedia/laravel-ffmpeg": "^7.5",
         "predis/predis": "^1.1",
         "spatie/flysystem-dropbox": "~1.2",
         "spatie/laravel-medialibrary": "^7.0.0"
     },
     "require-dev": {
-        "facade/ignition": "^2.3.6",
-        "fzaninotto/faker": "^1.9.1",
+        "barryvdh/laravel-debugbar": "^3.5",
+        "facade/ignition": "^2.5",
+        "fakerphp/faker": "^1.9.1",
+        "laravel/sail": "^1.0.1",
         "laravel/ui": "^3.0",
-        "mockery/mockery": "^1.3.1",
+        "mockery/mockery": "^1.4.2",
         "nunomaduro/collision": "^5.0",
-        "phpunit/phpunit": "^9.3"
+        "phpunit/phpunit": "^9.3.3"
     },
     "config": {
         "optimize-autoloader": true,
@@ -41,12 +45,10 @@
     },
     "autoload": {
         "psr-4": {
-            "Laratube\\": "app/"
-        },
-        "classmap": [
-            "database/seeds",
-            "database/factories"
-        ]
+            "App\\": "app/",
+            "Database\\Factories\\": "database/factories/",
+            "Database\\Seeders\\": "database/seeders/"
+        }
     },
     "autoload-dev": {
         "psr-4": {
