@@ -1,8 +1,9 @@
 <?php
 
-namespace Laratube\Http\Controllers\Auth;
+namespace App\Http\Controllers\Auth;
 
-use Laratube\Http\Controllers\Controller;
+use App\Http\Controllers\Controller;
+use App\Providers\RouteServiceProvider;
 use Illuminate\Foundation\Auth\VerifiesEmails;
 
 class VerificationController extends Controller
@@ -25,7 +26,7 @@ class VerificationController extends Controller
      *
      * @var string
      */
-    protected $redirectTo = '/home';
+    protected $redirectTo = RouteServiceProvider::HOME;
 
     /**
      * Create a new controller instance.
