@@ -1,12 +1,13 @@
 <?php
 
-namespace Laratube\Http\Controllers;
+namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
-use Laratube\Channel;
-use Laratube\Video;
-use Laratube\Jobs\Videos\ConvertForStreaming;
-use Laratube\Jobs\Videos\CreateVideoThumbnail;
+use App\Models\Channel;
+use App\Models\Video;
+use App\Jobs\Videos\ConvertForStreaming;
+use App\Jobs\Videos\CreateVideoThumbnail;
+use Illuminate\Support\Facades\Storage;
 
 class UploadVideoController extends Controller
 {
@@ -17,10 +18,12 @@ class UploadVideoController extends Controller
 
     public function store(Channel $channel, Request $request)
     {
+//        $path = $request->video->store("/FFMPEG/channels/{$channel->id}/videos", 'public');
 
         $video = $channel->videos()->create([
             'title' => $request->title,
-            'path' => $request->video->store("channels/{$channel->id}/videos")
+            'path' => $request->video->store("/FFMPEG/channels/{$channel->id}/videos")
+//            'path' => Storage::disk('public')->url($path)
         ]);
 
         $this->dispatch(new CreateVideoThumbnail($video));
