@@ -1,7 +1,8 @@
 <?php
 
-namespace Laratube;
+namespace App\Models;
 
+use App\Application\Uuids;
 use Illuminate\Notifications\Notifiable;
 use Illuminate\Contracts\Auth\MustVerifyEmail;
 use Illuminate\Foundation\Auth\User as Authenticatable;
@@ -9,7 +10,7 @@ use Illuminate\Support\Str;
 
 class User extends Authenticatable
 {
-    use Notifiable;
+    use Notifiable, Uuids;
 
     /**
      * The attributes that are mass assignable.
@@ -21,16 +22,6 @@ class User extends Authenticatable
     protected $fillable = [
         'name', 'email', 'password',
     ];
-
-    protected static function boot()
-    {
-        parent::boot();
-
-        static::creating(function ($model) {
-        // $model->id = Str::uuid();
-            $model->{$model->getKeyName()} = (string) Str::uuid();
-        });
-    }
 
     /**
      * The attributes that should be hidden for arrays.
