@@ -1,9 +1,9 @@
 <?php
 
-namespace Laratube\Providers;
+namespace App\Providers;
 
-use Illuminate\Support\Facades\Gate;
 use Illuminate\Foundation\Support\Providers\AuthServiceProvider as ServiceProvider;
+use Illuminate\Support\Facades\Gate;
 
 class AuthServiceProvider extends ServiceProvider
 {
@@ -13,7 +13,7 @@ class AuthServiceProvider extends ServiceProvider
      * @var array
      */
     protected $policies = [
-        'Laratube\Model' => 'Laratube\Policies\ModelPolicy',
+//         'App\Models\Model' => 'App\Policies\ModelPolicy',
     ];
 
     /**
