@@ -1,12 +1,12 @@
 <?php
 
-namespace Laratube\Providers;
+namespace App\Providers;
 
-use Illuminate\Support\Facades\Event;
+use App\Listeners\Users\CreateUserChannelListener;
 use Illuminate\Auth\Events\Registered;
 use Illuminate\Auth\Listeners\SendEmailVerificationNotification;
 use Illuminate\Foundation\Support\Providers\EventServiceProvider as ServiceProvider;
-use Laratube\Listeners\User\CreateUserChannel;
+use Illuminate\Support\Facades\Event;
 
 class EventServiceProvider extends ServiceProvider
 {
@@ -18,7 +18,7 @@ class EventServiceProvider extends ServiceProvider
     protected $listen = [
         Registered::class => [
             SendEmailVerificationNotification::class,
-            CreateUserChannel::class
+            CreateUserChannelListener::class
         ],
     ];
 
@@ -29,8 +29,6 @@ class EventServiceProvider extends ServiceProvider
      */
     public function boot()
     {
-        parent::boot();
-
         //
     }
 }
