@@ -4,57 +4,33 @@ return [
 
     /*
     |--------------------------------------------------------------------------
-    | Laravel CORS Options
+    | Cross-Origin Resource Sharing (CORS) Configuration
     |--------------------------------------------------------------------------
     |
-    | The allowed_methods and allowed_headers options are case-insensitive.
+    | Here you may configure your settings for cross-origin resource sharing
+    | or "CORS". This determines what cross-origin operations may execute
+    | in web browsers. You are free to adjust these settings as needed.
     |
-    | You don't need to provide both allowed_origins and allowed_origins_patterns.
-    | If one of the strings passed matches, it is considered a valid origin.
-    |
-    | If array('*') is provided to allowed_methods, allowed_origins or allowed_headers
-    | all methods / origins / headers are allowed.
+    | To learn more: https://developer.mozilla.org/en-US/docs/Web/HTTP/CORS
     |
     */
 
-    /*
-     * You can enable CORS for 1 or multiple paths.
-     * Example: ['api/*']
-     */
-    'paths' => [],
+    'paths' => ['api/*', 'sanctum/csrf-cookie', '*'],
 
-    /*
-    * Matches the request method. `[*]` allows all methods.
-    */
     'allowed_methods' => ['*'],
 
-    /*
-     * Matches the request origin. `[*]` allows all origins.
-     */
     'allowed_origins' => ['*'],
 
-    /*
-     * Matches the request origin with, similar to `Request::is()`
-     */
     'allowed_origins_patterns' => [],
 
-    /*
-     * Sets the Access-Control-Allow-Headers response header. `[*]` allows all headers.
-     */
     'allowed_headers' => ['*'],
 
-    /*
-     * Sets the Access-Control-Expose-Headers response header with these headers.
-     */
     'exposed_headers' => [],
 
-    /*
-     * Sets the Access-Control-Max-Age response header when > 0.
-     */
     'max_age' => 0,
 
-    /*
-     * Sets the Access-Control-Allow-Credentials header.
-     */
     'supports_credentials' => false,
+
+
+
 ];
