@@ -1,12 +1,6 @@
 <?php
 
-
-$url = parse_url(getenv("CLEARDB_DATABASE_URL"));
-
-$host = $url["host"] ?? '';
-$username = $url["user"] ?? '';
-$password = $url["pass"] ?? '';
-$database = substr($url["path"], 1) ?? '';
+use Illuminate\Support\Str;
 
 return [
 
@@ -21,8 +15,7 @@ return [
     |
     */
 
-    // 'default' => env('DB_CONNECTION', 'heroku_mysql'),
-	'default' => 'mysql',
+    'default' => env('DB_CONNECTION', 'mysql'),
 
     /*
     |--------------------------------------------------------------------------
@@ -44,43 +37,35 @@ return [
 
         'sqlite' => [
             'driver' => 'sqlite',
+            'url' => env('DATABASE_URL'),
             'database' => env('DB_DATABASE', database_path('database.sqlite')),
             'prefix' => '',
             'foreign_key_constraints' => env('DB_FOREIGN_KEYS', true),
         ],
 
-         'mysql' => [
-             'driver' => 'mysql',
-             'host' => env('db_host', '127.0.0.1'),
-             'port' => env('db_port', '3306'),
-             'database' => env('DB_DATABASE', 'forge'),
-             'username' => env('DB_USERNAME', 'forge'),
-             'password' => env('DB_PASSWORD', ''),
-             'unix_socket' => env('db_socket', ''),
-             'charset' => 'utf8mb4',
-             'collation' => 'utf8mb4_unicode_ci',
-             'prefix' => '',
-             'prefix_indexes' => true,
-             'strict' => true,
-             'engine' => null,
-//             'options' => array_filter([
-                 // pdo::mysql_attr_ssl_ca => env('mysql_attr_ssl_ca'),
-//             ]),
-         ],
-
-        'heroku_mysql' => [
+        'mysql' => [
             'driver' => 'mysql',
-            'host' => $host,
-            'database' => $database,
-            'username' => $username,
-            'password' => $password,
-            'charset' => 'utf8',
-            'collation' => 'utf8_unicode_ci',
+            'url' => env('DATABASE_URL'),
+            'host' => env('DB_HOST', '127.0.0.1'),
+            'port' => env('DB_PORT', '3306'),
+            'database' => env('DB_DATABASE', 'forge'),
+            'username' => env('DB_USERNAME', 'forge'),
+            'password' => env('DB_PASSWORD', ''),
+            'unix_socket' => env('DB_SOCKET', ''),
+            'charset' => 'utf8mb4',
+            'collation' => 'utf8mb4_unicode_ci',
             'prefix' => '',
+            'prefix_indexes' => true,
+            'strict' => true,
+            'engine' => null,
+            'options' => extension_loaded('pdo_mysql') ? array_filter([
+                PDO::MYSQL_ATTR_SSL_CA => env('MYSQL_ATTR_SSL_CA'),
+            ]) : [],
         ],
 
         'pgsql' => [
             'driver' => 'pgsql',
+            'url' => env('DATABASE_URL'),
             'host' => env('DB_HOST', '127.0.0.1'),
             'port' => env('DB_PORT', '5432'),
             'database' => env('DB_DATABASE', 'forge'),
@@ -95,6 +80,7 @@ return [
 
         'sqlsrv' => [
             'driver' => 'sqlsrv',
+            'url' => env('DATABASE_URL'),
             'host' => env('DB_HOST', 'localhost'),
             'port' => env('DB_PORT', '1433'),
             'database' => env('DB_DATABASE', 'forge'),
@@ -131,28 +117,31 @@ return [
     |
     */
 
-//    'redis' => [
-//
-//        'client' => env('REDIS_CLIENT', 'predis'),
-//
-//        'options' => [
-//            'cluster' => env('REDIS_CLUSTER', 'predis'),
-//        ],
-//
-//        'default' => [
-//            'host' => env('REDIS_HOST', '127.0.0.1'),
-//            'password' => env('REDIS_PASSWORD', null),
-//            'port' => env('REDIS_PORT', 6379),
-//            'database' => env('REDIS_DB', 0),
-//        ],
-//
-//        'cache' => [
-//            'host' => env('REDIS_HOST', '127.0.0.1'),
-//            'password' => env('REDIS_PASSWORD', null),
-//            'port' => env('REDIS_PORT', 6379),
-//            'database' => env('REDIS_CACHE_DB', 1),
-//        ],
-//
-//    ],
+    'redis' => [
+
+        'client' => env('REDIS_CLIENT', 'phpredis'),
+
+        'options' => [
+            'cluster' => env('REDIS_CLUSTER', 'redis'),
+            'prefix' => env('REDIS_PREFIX', Str::slug(env('APP_NAME', 'laravel'), '_').'_database_'),
+        ],
+
+        'default' => [
+            'url' => env('REDIS_URL'),
+            'host' => env('REDIS_HOST', '127.0.0.1'),
+            'password' => env('REDIS_PASSWORD', null),
+            'port' => env('REDIS_PORT', '6379'),
+            'database' => env('REDIS_DB', '0'),
+        ],
+
+        'cache' => [
+            'url' => env('REDIS_URL'),
+            'host' => env('REDIS_HOST', '127.0.0.1'),
+            'password' => env('REDIS_PASSWORD', null),
+            'port' => env('REDIS_PORT', '6379'),
+            'database' => env('REDIS_CACHE_DB', '1'),
+        ],
+
+    ],
 
 ];
