@@ -11,12 +11,6 @@ const mix = require('laravel-mix');
  |
  */
 
-mix.webpackConfig({
-    output: {
-        publicPath: '/',
-        chunkFilename: 'js/[name].[chunkhash].js',
-    }
-});
-
 mix.js('resources/js/app.js', 'public/js')
-   .sass('resources/sass/app.scss', 'public/css');
+    .vue()
+    // .sass('resources/sass/app.scss', 'public/css');
